__template__/requirements.txt
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
db-dtypes==1.1.1
et-xmlfile==1.1.0
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.11.3
google-cloud-core==2.3.2
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
gspread==5.10.0
httplib2==0.22.0
idna==3.4
numpy==1.25.0
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
proto-plus==1.22.3
protobuf==4.23.3
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
